backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.133.0",
    "httpx>=0.28.1",
    "numpy>=1.26.0",
    "pymupdf>=1.27.1",
    "python-pptx>=1.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "groq>=0.25.0",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
